% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrappind function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of sample iterations}

\item{x}{sample}

\item{fun}{statistic}

\item{alpha}{alpha parameter to determine confidence interval}

\item{cx}{text character expansion factor}

\item{...}{extra arguments to be passed to histogram function}
}
\value{
list containing the confidence interval, statistic, and sample
}
\description{
Creates a point and interval estimate using the bootstrap method of a given sample and statistic
}
