% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{95\% Confidence interval}
\usage{
myci(x, conf.level = 0.95)
}
\arguments{
\item{x}{sample}
}
\description{
\code{myci} creates a 95\% confidence interval of the mean from a single sample
}
