% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Number of tickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{the number of seats}

\item{gamma}{the probability of an overbooking}

\item{p}{the probability of a "no-show"}
}
\value{
A list containing nd and nc, the discrete and continuous number of tickets that should be sold, respectively, and N, gamma, and p.
}
\description{
\code{ntickets} calculates how many tickets should be sold to ovoid overbooking
}
\details{
This function takes the number of total seats, the desired probability of a booking, and the pobability of a "no-show", and provides a discrete and coninuous number of tickets that should be sold to avoid overbooking. It plots the objective function vs n as well, both continuous and discrete.
}
