% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltp.R
\name{mycltp}
\alias{mycltp}
\title{Poisson central limit theorem}
\usage{
mycltp(n, iter = 10000, lambda = 4, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{lambda}{lambda of the poisson}

\item{...}{additional arguments to be given to the histogram}
}
\description{
\code{mycltp} plots the sample mean of the poisson distribution
}
